Pull up redirect and itemCollection to Transaction.php
